requirements.txt
black==22.3.0
DateTime==5.1
flake8==4.0.1
isort==5.10.1
mypy==v0.931
mysql-connector-python==8.0.33
pandas==2.0.2
pymysql==1.0.3
